drop old scala library, mist logger is useless after #462
