+++ examples/examples/src/main/scala/SparkContextExample.scala
@@ -2,8 +2,9 @@ import mist.api._
 import mist.api.MistExtras
 import mist.api.encoding.DefaultEncoders._
 import org.apache.spark.SparkContext
+import org.apache.spark.internal.Logging
 
-object SparkContextExample extends MistFn[Array[Int]] {
+object SparkContextExample extends MistFn[Array[Int]] with Logging {
 
   override def handle = {
     withArgs(
@@ -14,7 +15,7 @@ object SparkContextExample extends MistFn[Array[Int]] {
 
        import extras._
 
-       logger.info(s"Heello from $jobId")
+       logInfo(s"Heello from $jobId")
        sc.parallelize(nums).map(_ * mult).collect()
      })
   }
