--- mist-lib/src/main/scala/io/hydrosphere/mist/api/MistJob.scala
@@ -1,3 +1,0 @@
-package io.hydrosphere.mist.api
-
-trait MistJob extends ContextSupport
