+++ mist-lib/src/main/scala/mist/api/FnContext.scala
@@ -1,21 +1,29 @@
 package mist.api
 
-import io.hydrosphere.mist.api.SetupConfiguration
+import org.apache.spark.SparkContext
+import org.apache.spark.streaming.Duration
 
 sealed trait FnContext{
   val params: Map[String, Any]
 }
 
 case class FullFnContext(
-  setupConf: SetupConfiguration,
+  sc: SparkContext,
+  streamingDuration: Duration,
+  info: RuntimeJobInfo,
   params: Map[String, Any]
 ) extends FnContext
 
 object FnContext {
+
   def apply(userParams: Map[String, Any]): FnContext = new FnContext {
     override val params: Map[String, Any] = userParams
   }
 
-  def apply(setupConf: SetupConfiguration, params: Map[String, Any]) =
-    FullFnContext(setupConf, params)
+  def apply(
+    sc: SparkContext,
+    params: Map[String, Any],
+    streamingDuration: Duration = Duration(1000),
+    info: RuntimeJobInfo = RuntimeJobInfo.Unknown): FullFnContext = FullFnContext(sc, streamingDuration, info, params)
+
 }
