+++ mist-lib/src/main/scala/mist/api/RuntimeJobInfo.scala
@@ -1,0 +1,10 @@
+package mist.api
+
+case class RuntimeJobInfo(
+  id: String,
+  workerId: String
+)
+
+object RuntimeJobInfo {
+  val Unknown = RuntimeJobInfo("unknown", "unknown")
+}
