+++ mist-lib/src/main/scala/mist/api/package.scala
@@ -2,6 +2,4 @@ package mist
 
 package object api extends Contexts {
 
-  type MLogger = io.hydrosphere.mist.api.MLogger
-
 }
