+++ mist/core/src/main/scala/io/hydrosphere/mist/core/jvmjob/FunctionInstanceLoader.scala
@@ -4,10 +4,8 @@ import java.io.File
 import java.net.URLClassLoader
 
 import io.hydrosphere.mist.core.CommonData.Action
-import io.hydrosphere.mist.utils.{Err, TryLoad, Succ}
+import io.hydrosphere.mist.utils.{Err, TryLoad}
 import mist.api.internal.BaseFunctionInstance
-
-import scala.reflect.runtime.universe._
 
 class FunctionInstanceLoader(val classLoader: ClassLoader) {
 
@@ -18,25 +16,9 @@ class FunctionInstanceLoader(val classLoader: ClassLoader) {
       case clz if mist.api.internal.FunctionInstance.isJavaInstance(clz) =>
         TryLoad(mist.api.internal.FunctionInstance.loadJava(clz))
       case clz =>
-        loadFnInstance(clz, action) match {
-          case Some(i) => Succ(i)
-          case None =>
-            val e = new IllegalStateException(s"Can not instantiate job for action $action")
-            Err(e)
-        }
+        val e = new IllegalStateException(s"Can not instantiate job for action $action")
+        Err(e)
     })
-  }
-
-  private def loadFnInstance(clazz: Class[_], action: Action): Option[OldInstanceWrapper] = {
-    val methodName = methodNameByAction(action)
-    val term = newTermName(methodName)
-    val symbol = runtimeMirror(clazz.getClassLoader).classSymbol(clazz).toType.member(term)
-    if (!symbol.isMethod) {
-      None
-    } else {
-      val instance = new FunctionInstance(clazz, symbol.asMethod)
-      Some(new OldInstanceWrapper(instance))
-    }
   }
 
   private def methodNameByAction(action: Action): String = action match {
