--- mist/core/src/test/scala/io/hydrosphere/mist/core/jvmjob/testJobs.scala
@@ -1,51 +1,0 @@
-package io.hydrosphere.mist.core.jvmjob
-
-import io.hydrosphere.mist.api._
-
-object InvalidJob extends MistJob
-
-object MultiplyJob extends MistJob {
-
-  def execute(numbers: List[Int]): Map[String, Any] = {
-    val rdd = context.parallelize(numbers)
-    Map("r" -> rdd.map(x => x * 2).collect().toList)
-  }
-}
-
-object OptParamJob extends MistJob {
-
-  def execute(p: Option[Int]): Map[String, Any] = {
-    Map("r" -> p.getOrElse(42))
-  }
-}
-
-object ManyArgJob extends MistJob {
-
-  def execute(
-    FromDate: String,
-    ToDate: String,
-    query: String,
-    rows: Int,
-    Separator: String
-  ): Map[String, Any] = {
-
-    val expected = Seq("FromDate", "ToDate", "query", 1, "Separator")
-    val given = Seq(FromDate, ToDate, query, rows, Separator)
-
-    val isOk = expected == given
-    Map("isOk" -> isOk)
-  }
-}
-
-object AllMixinsJob extends MistJob
-  with StreamingSupport
-  with SQLSupport
-  with HiveSupport {
-
-  def execute(): Map[String, Any] = {
-    val ssc = streamingContext
-    ssc.start()
-    ssc.stop(stopSparkContext = false)
-    Map("result" -> "ok")
-  }
-}
