+++ mist/master/src/main/scala/io/hydrosphere/mist/master/interfaces/jsonCodecs.scala
@@ -3,8 +3,8 @@ package io.hydrosphere.mist.master.interfaces
 import java.time.LocalDateTime
 import java.time.format.{DateTimeFormatter, DateTimeParseException}
 
-import io.hydrosphere.mist.api.logging.MistLogging.LogEvent
 import io.hydrosphere.mist.core.CommonData.{Action, JobParams, WorkerInitInfo}
+import io.hydrosphere.mist.core.logging.LogEvent
 import io.hydrosphere.mist.master.Messages.StatusMessages._
 import io.hydrosphere.mist.master.execution.{WorkerFullInfo, WorkerLink}
 import io.hydrosphere.mist.master.interfaces.http._
