+++ mist/master/src/main/scala/io/hydrosphere/mist/master/logging/JobLogger.scala
@@ -1,7 +1,7 @@
 package io.hydrosphere.mist.master.logging
 
 import akka.actor.ActorRef
-import io.hydrosphere.mist.api.logging.MistLogging.LogEvent
+import io.hydrosphere.mist.core.logging.LogEvent
 
 trait JobLogger {
   protected val jobId: String
