+++ mist/master/src/main/scala/io/hydrosphere/mist/master/logging/LogStreams.scala
@@ -8,7 +8,7 @@ import akka.stream.{ActorAttributes, ActorMaterializer, OverflowStrategy}
 import akka.util.ByteString
 import akka.{Done, NotUsed}
 import com.twitter.chill.{KryoPool, ScalaKryoInstantiator}
-import io.hydrosphere.mist.api.logging.MistLogging.LogEvent
+import io.hydrosphere.mist.core.logging.LogEvent
 import io.hydrosphere.mist.master.Messages.StatusMessages.ReceivedLogs
 import io.hydrosphere.mist.master.{EventsStreamer, LogStoragePaths}
 import io.hydrosphere.mist.utils.Logger
