+++ mist/worker/src/main/scala/io/hydrosphere/mist/worker/logging/RemoteAppender.scala
@@ -1,8 +1,6 @@
-package io.hydrosphere.mist.worker
+package io.hydrosphere.mist.worker.logging
 
-import io.hydrosphere.mist.api.CentralLoggingConf
-import io.hydrosphere.mist.api.logging.MistLogging
-import io.hydrosphere.mist.api.logging.MistLogging.{LogsWriter, RemoteLogsWriter}
+import io.hydrosphere.mist.core.logging.LogEvent
 import org.apache.log4j.spi.LoggingEvent
 import org.apache.log4j.{AppenderSkeleton, Level, SimpleLayout}
 
@@ -13,16 +11,16 @@ class RemoteAppender(sourceId: String, logsWriter: LogsWriter) extends AppenderS
     val timeStamp = event.timeStamp
     val message = event.getRenderedMessage
     val evt = event.getLevel match {
-      case Level.INFO => MistLogging.LogEvent.mkInfo(sourceId, message, timeStamp)
-      case Level.DEBUG => MistLogging.LogEvent.mkDebug(sourceId, message, timeStamp)
+      case Level.INFO => LogEvent.mkInfo(sourceId, message, timeStamp)
+      case Level.DEBUG => LogEvent.mkDebug(sourceId, message, timeStamp)
       case Level.ERROR =>
-        MistLogging.LogEvent.mkError(
+        LogEvent.mkError(
           sourceId, message,
           Option(event.getThrowableInformation).map(_.getThrowable),
           timeStamp
         )
-      case Level.WARN => MistLogging.LogEvent.mkWarn(sourceId, message, timeStamp)
-      case _ => MistLogging.LogEvent.mkInfo(sourceId, this.getLayout.format(event), timeStamp)
+      case Level.WARN => LogEvent.mkWarn(sourceId, message, timeStamp)
+      case _ => LogEvent.mkInfo(sourceId, this.getLayout.format(event), timeStamp)
     }
     logsWriter.write(evt)
   }
@@ -34,8 +32,6 @@ class RemoteAppender(sourceId: String, logsWriter: LogsWriter) extends AppenderS
 
 
 object RemoteAppender {
-  def apply(sourceId: String, loggingConf: CentralLoggingConf): RemoteAppender =
-    create(sourceId, RemoteLogsWriter.getOrCreate(loggingConf.host, loggingConf.port))
 
   def create(sourceId: String, logsWriter: LogsWriter): RemoteAppender = {
     val jobLogsAppender = new RemoteAppender(sourceId, logsWriter)
