+++ mist/worker/src/main/scala/io/hydrosphere/mist/worker/runners/JobRunner.scala
@@ -1,14 +1,14 @@
 package io.hydrosphere.mist.worker.runners
 
 import io.hydrosphere.mist.core.CommonData.RunJobRequest
-import io.hydrosphere.mist.worker.NamedContext
+import io.hydrosphere.mist.worker.MistScContext
 import mist.api.data.JsLikeData
 
 import scala.util._
 
 trait JobRunner {
 
-  def run(req: RunJobRequest, context: NamedContext): Either[Throwable, JsLikeData]
+  def run(req: RunJobRequest, context: MistScContext): Either[Throwable, JsLikeData]
 }
