+++ mist/worker/src/main/scala/io/hydrosphere/mist/worker/runners/python/wrappers/SparkStreamingWrapper.scala
@@ -1,15 +1,9 @@
 package io.hydrosphere.mist.worker.runners.python.wrappers
 
-import io.hydrosphere.mist.api._
-import org.apache.spark.streaming.api.java.JavaStreamingContext
+import org.apache.spark.streaming.Duration
 
-class SparkStreamingWrapper(config: SetupConfiguration) extends StreamingSupport {
+//TODO remove
+class SparkStreamingWrapper(duration: Duration) {
 
-  _context = config.context
-  
-  def setStreamingContext(ssc: JavaStreamingContext): Unit = {
-    _ssc = ssc.ssc
-  }
-
-  def getDurationSeconds: Int = (config.streamingDuration.milliseconds / 1000).toInt
+  def getDurationSeconds: Int = (duration.milliseconds / 1000).toInt
 }
